Large Update
This is a large update that hopes to polish the app a lot.  Colors have been added and changed. Section scaling has been changed. Bugs have been fixed. Mission and Project Icons have changed. All content windows have changed with new graphics and tools directly inside the text window. The tutorial was populated with an actual, working, tutorial.

One quality of life thing is that when the user clicks on the panel, the apps panel zooms to the center of the screen. Especially for when we scroll past it and stuff, it always automatically returns itself to the center.

It is still not complete, nor perfect, but it is so much better.