censoring data part 1
I am censoring the dataset. This is an act to protect researchers from losing their jobs or their funding, although is a rather terrible thing to do. It feels like burning a pamphlet or directory, more than burning a book, but is close enough.

When the time comes, it will be uncensored. Otherwise, reach out with questions comments or concerns. The data is still available, just not shown as directly on GH-Hub.